+++ src/DevilDaggersInfo.Core.Mod/ModBinaryTocEntry.cs
@@ -8,4 +8,14 @@ public record ModBinaryTocEntry(string Name, int Offset, int Size, AssetType Ass
 	{
 		return AssetType == AssetType.Audio && Name == "loudness";
 	}
+
+	public ModBinaryTocEntry Enable()
+	{
+		return this with { Name = Name.ToLower() };
+	}
+
+	public ModBinaryTocEntry Disable()
+	{
+		return this with { Name = Name.ToUpper() };
+	}
 }